+++ pyjoulescope_driver/__init__.py
@@ -14,12 +14,13 @@
 
 from .version import *
 from . import time64
+from .record import Record
 import os
 if os.environ.get('pyjoulescope_driver_setup', '0') != '1':
     from .binding import Driver
 
 __all__ = [
-    'Driver',
+    'Driver', 'Record',
     'time64',
     '__version__', '__title__', '__description__', '__url__',
     '__author__', '__author_email__', '__license__',